refactor(all): 파일 구조를 폴더로 정리
