+++ src/functions/getInput.js
