+++ src/functions/printResult.js
